+++ Exodus/src/utopia/exodus/util/ExodusContext.scala
@@ -75,7 +75,8 @@ object ExodusContext
 	@throws[EnvironmentNotSetupException]("If .setup(...) hasn't been called yet")
 	private def get = data match {
 		case Some(data) => data
-		case None => throw EnvironmentNotSetupException("ExodusContext.setup must be called before using this method")
+		case None =>
+			throw new EnvironmentNotSetupException("ExodusContext.setup must be called before using this method")
 	}