+++ Vigil/src/utopia/vigil/database/store/ScopeDb.scala
@@ -1,10 +1,13 @@
 package utopia.vigil.database.store
 
+import utopia.flow.collection.immutable.{Empty, OptimizedIndexedSeq, Tree}
 import utopia.vault.database.{Connection, Store}
-import utopia.vigil.database.access.scope.AccessScope
+import utopia.vault.store.StoreResult
+import utopia.vigil.database.access.scope.{AccessScope, AccessScopes}
 import utopia.vigil.database.storable.scope.ScopeDbModel
 import utopia.vigil.model.cached.scope.ScopeTarget
 import utopia.vigil.model.partial.scope.ScopeData
+import utopia.vigil.model.stored.scope.Scope
 
 /**
  * Used for interacting with scope information in the DB
@@ -48,4 +51,35 @@ object ScopeDb
 	 */
 	def store(scope: String, parent: Option[ScopeTarget] = None)(implicit connection: Connection) =
 		_store.single(scope -> parent.filter { _.isValid }.map { _.id }, AccessScope.forKey(scope).pull)
+	/**
+	 * Stores scope trees to the database
+	 * @param scopeTrees Scope trees to store, where navs are scope keys.
+	 * @param connection Implicit DB connection
+	 * @return Scope store results
+	 */
+	def store(scopeTrees: Iterable[Tree[String]])(implicit connection: Connection) = {
+		if (scopeTrees.nonEmpty) {
+			// Pulls the existing scope info in order to avoid duplicates
+			val existing = AccessScopes.pull
+			val resultBuilder = OptimizedIndexedSeq.newBuilder[StoreResult[Scope]]
+			
+			// Stores one scope layer at a time
+			var nextScopes = scopeTrees.map[(Tree[String], Option[Int])] { _ -> None }
+			while (nextScopes.nonEmpty) {
+				val stored = _store.keyMap(nextScopes.map { case (node, parentId) => node.nav -> parentId }, existing) {
+					_._1.toLowerCase } { _.key.toLowerCase }
+				
+				resultBuilder ++= stored.valuesIterator
+				nextScopes = nextScopes.flatMap { case (node, _) =>
+					val parentId = stored(node.nav.toLowerCase).id
+					node.children.map { _ -> Some(parentId) }
+				}
+			}
+			
+			resultBuilder.result()
+		}
+		// Case: Nothing to store
+		else
+			Empty
+	}
 }